ai_text_image/requirements.txt
# Core dependencies (Python 3.8+)
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pydantic>=2.0.0
requests>=2.28.0

# MCP Server (requires Python 3.10+)
# Install separately if using Python 3.10+: pip install mcp>=1.0.0
# mcp>=1.0.0

